refactor: use int16_t for temperature alarm thresholds instead of float
Storing thresholds as float was unnecessary: all defaults are whole-number
degrees (50, 70, 70, 81), Firefly temps arrive as int8_t, and the +5 deg C
overtemp margin makes sub-degree precision irrelevant. Float storage also
required memcpy for safe type-punning and strtof for CLI parsing, both of
which pull in library code that is costly on a space-constrained MCU.

Changes:
- alarmTemp[]: float -> int16_t
- INITIAL_ALARM_TEMP_* constants: remove .0f suffix
- ALM_OVERTEMP_THRESHOLD: 5.0f -> 5
- setAlarmTemperature(): int16_t parameter; EEPROM write zero-extends
  uint16_t to uint32_t (sentinel 0xFFFFFFFF remains safely distinct)
- loadAlarmTemperaturesFromEEPROM(): cast uint16_t back to int16_t on read
- Remove #include <string.h> (no longer needed)
- TempStatus(): no change needed; int16_t promotes to float implicitly
  in the float subtraction against currentTemp[]
- alm status CLI: replace float_to_ints() calls with direct %d formatting
- alm settemp CLI: revert strtof back to strtol

+++ projects/cm_mcu/Tasks.h
@@ -206,8 +206,8 @@ enum powdevice { GEN,
 
 void GenericAlarmTask(void *parameters);
 
-float getAlarmTemperature(enum device device_name);
-void setAlarmTemperature(enum device device_name, float newtemp);
+int16_t getAlarmTemperature(enum device device_name);
+void setAlarmTemperature(enum device device_name, int16_t newtemp);
 uint32_t getTempAlarmStatus(void);
 float getAlarmVoltageThres(void);
 void setAlarmVoltageThres(float voltthres);
